requirements.txt
aiogram
aiohttp
aioredis
python-dotenv
gino